plugin updates

+++ config/.config/fish/functions/nvm.fish
@@ -11,16 +11,16 @@ function nvm -a cmd v -d "Node version manager"
 
     switch "$cmd"
         case -v --version
-            echo "nvm, version 2.0.0"
+            echo "nvm, version 2.0.1"
         case "" -h --help
-            echo "Usage: nvm install <version>    Download and activate a given version"
+            echo "Usage: nvm install <version>    Download and activate the specified Node version"
             echo "       nvm install              Install version from nearest .nvmrc file"
             echo "       nvm use <version>        Activate a version in the current shell"
             echo "       nvm use                  Activate version from nearest .nvmrc file"
             echo "       nvm list                 List installed versions"
             echo "       nvm list-remote          List versions available to install"
             echo "       nvm list-remote <regex>  List versions matching a given regular expression"
-            echo "       nvm current              Print currently-active version"
+            echo "       nvm current              Print the currently-active version"
             echo "       nvm uninstall <version>  Uninstall a version"
             echo "Options:"
             echo "       -v or --version          Print version"
@@ -39,33 +39,39 @@ function nvm -a cmd v -d "Node version manager"
             end
 
             if test ! -e $nvm_data/$v
+                set --local os (uname -s | string lower)
+                set --local ext tar.gz
                 set --local arch (uname -m)
-                set --local os (string lower (uname -s))
 
                 switch $os
+                    case aix
+                        set arch ppc64
+                    case sunos
                     case linux
-                        switch $arch
-                            case x86_64
-                                set arch x64
-                            case armv6 armv6l
-                                set arch armv6l
-                            case armv7 armv7l
-                                set arch armv7l
-                            case armv8 armv8l aarch64
-                                set arch arm64
-                            case \*
-                                echo "nvm: Unsupported hardware architecture: \"$arch\"" >&2
-                                return 1
-                        end
                     case darwin
-                        set arch x64
+                    case {MSYS_NT,MINGW\*_NT}\*
+                        set os win
+                        set ext zip
                     case \*
                         echo "nvm: Unsupported operating system: \"$os\"" >&2
                         return 1
                 end
 
+                switch $arch
+                    case i\*86
+                        set arch x86
+                    case x86_64
+                        set arch x64
+                    case armv6 armv6l
+                        set arch armv6l
+                    case armv7 armv7l
+                        set arch armv7l
+                    case armv8 armv8l aarch64
+                        set arch arm64
+                end
+
                 set --local dir "node-$v-$os-$arch"
-                set --local url $nvm_mirror/$v/$dir.tar.gz
+                set --local url $nvm_mirror/$v/$dir.$ext
 
                 command mkdir -p $nvm_data/$v
 
@@ -81,8 +87,12 @@ function nvm -a cmd v -d "Node version manager"
 
                 echo -en "\033[F\33[2K\x1b[0m"
 
-                command mv $nvm_data/$v/$dir/* $nvm_data/$v
-                command rm -rf $nvm_data/$v/$dir
+                if test "$os" = "win"
+                    command mv $nvm_data/$v/$dir $nvm_data/$v/bin
+                else
+                    command mv $nvm_data/$v/$dir/* $nvm_data/$v
+                    command rm -rf $nvm_data/$v/$dir
+                end
             end
 
             if test $v != "$nvm_current_version"
